whatsapp-chatbot/services/flask_services/requirements.txt
flask==2.0.1
werkzeug==2.0.1
spacy==3.7.2 
python-dotenv==1.0.1
gunicorn==20.1.0